requirements.txt
flask
flask_mysqldb
mysql
mysql-connector
mysqlclient
bcrypt
gunicorn
Flask-MySQLdb==1.0.1
Flask-SQLAlchemy==2.5.1
Flask-JWT-Extended==4.3.1
Flask-RESTful==0.3.9
python-dotenv==0.21.0
bcrypt==4.0.1